finished moving components over to actual folder
