+++ components/contact/EmailForm.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import axios from "axios";
 import {
   useState,
