+++ components/contact/Toast.tsx
@@ -1,5 +1,4 @@
-"use client";
-import { useDarkMode } from "../../../components/providers/DarkModeProvider";
+import { useDarkMode } from "@/components/providers/DarkModeProvider";
 import { ToastContainer, Bounce } from "react-toastify";
 
 const Toast = () => {
